receipt done, changing to serializable

+++ Receipt.java
@@ -1,5 +1,4 @@
-
 public interface Receipt
 {
-   public double getTotalPrice();
+    public int getTotalPrice();
 }
